refactor: move messages to util folder
